Made TaskHelpers internal members
