+++ src/CacheCow.Client/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Runtime.CompilerServices;
 
 [assembly: AssemblyTitle("CacheCow.Client")]
 [assembly: AssemblyDescription("Client library for CacheCow project")]
-[assembly: InternalsVisibleTo("CacheCow.Tests")]
+[assembly: InternalsVisibleTo("CacheCow.Client.Tests")]
