+++ src/CacheCow.Common/CacheCow.Common.csproj
@@ -64,8 +64,6 @@
     <Compile Include="Helpers\BasicExtensions.cs" />
     <Compile Include="Helpers\HttpHeadersExtensions.cs" />
     <Compile Include="Helpers\HttpResponseMessageExtensions.cs" />
-    <Compile Include="Helpers\TaskHelpers.cs" />
-    <Compile Include="Helpers\TaskHelpersExtensions.cs" />
     <Compile Include="HttpHeaderNames.cs" />
     <Compile Include="Http\NotModifiedResponse.cs" />
     <Compile Include="ICacheStore.cs" />
