+++ test/CacheCow.Client.Tests/CacheCow.Client.Tests.csproj
@@ -1,15 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProductVersion>8.0.30703</ProductVersion>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{B67727F4-5B65-41D0-8D26-8D1C32CEA63D}</ProjectGuid>
+    <ProjectGuid>{1ADB6175-95B9-4BB9-AF56-B4291196554C}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>CacheCow.Server</RootNamespace>
-    <AssemblyName>CacheCow.Server</AssemblyName>
+    <RootNamespace>CacheCow.Client.Tests</RootNamespace>
+    <AssemblyName>CacheCow.Client.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\..\</SolutionDir>
@@ -33,8 +32,14 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Newtonsoft.Json, Version=4.5.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Newtonsoft.Json.4.5.10\lib\net40\Newtonsoft.Json.dll</HintPath>
+    <Reference Include="Newtonsoft.Json">
+      <HintPath>..\..\packages\Newtonsoft.Json.4.5.11\lib\net40\Newtonsoft.Json.dll</HintPath>
+    </Reference>
+    <Reference Include="nunit.framework">
+      <HintPath>..\..\packages\NUnit.2.6.2\lib\nunit.framework.dll</HintPath>
+    </Reference>
+    <Reference Include="Rhino.Mocks">
+      <HintPath>..\..\packages\RhinoMocks.3.6.1\lib\net\Rhino.Mocks.dll</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -54,29 +59,38 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Common\CommonAssemblyInfo.cs">
-      <Link>Properties\CommonAssemblyInfo.cs</Link>
+    <Compile Include="..\Common\DummyMessageHandler.cs">
+      <Link>Common\DummyMessageHandler.cs</Link>
     </Compile>
-    <Compile Include="..\Common\TaskHelpers.cs">
-      <Link>Common\TaskHelpers.cs</Link>
-    </Compile>
-    <Compile Include="..\Common\TaskHelpersExtensions.cs">
-      <Link>Common\TaskHelpersExtensions.cs</Link>
-    </Compile>
-    <Compile Include="CachingHandler.cs" />
-    <Compile Include="InMemoryEntityTagStore.cs" />
+    <Compile Include="CacheCowHeaderTests.cs" />
+    <Compile Include="CacheStoreQuotaManagerTests.cs" />
+    <Compile Include="CachingHandlerTests.cs" />
+    <Compile Include="Helper\DeepComparer.cs" />
+    <Compile Include="Helper\ResponseHelper.cs" />
+    <Compile Include="IntegrationTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RequestSerializationTests.cs" />
+    <Compile Include="ResponseSerializationTests.cs" />
+    <Compile Include="ResponseValidatorTests.cs" />
+    <Compile Include="SerialisationTests.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\CacheCow.Common\CacheCow.Common.csproj">
+    <ProjectReference Include="..\..\src\CacheCow.Client\CacheCow.Client.csproj">
+      <Project>{943325cb-7189-42fe-9da4-e513db3e0e87}</Project>
+      <Name>CacheCow.Client</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\src\CacheCow.Common\CacheCow.Common.csproj">
       <Project>{88e170e4-b857-42fe-963c-0505502b548c}</Project>
       <Name>CacheCow.Common</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="packages.config">
-      <SubType>Designer</SubType>
-    </None>
+    <EmbeddedResource Include="Data\Request.bin" />
+    <EmbeddedResource Include="Data\Response.bin" />
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Helper\Headers.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
