+++ test/CacheCow.Client.Tests/Data/Response.bin
