+++ test/CacheCow.Client.Tests/Helper/Headers.txt
@@ -1,0 +1,11 @@
+﻿﻿date: Tue, 24 Jul 2012 10:23:56 GMT
+content-encoding: gzip
+status: 200 OK
+content-length: 27603
+x-xss-protection: 1; mode=block
+server: gws
+x-frame-options: SAMEORIGIN
+content-type: text/html; charset=UTF-8
+expires: -1
+cache-control: private, max-age=0
+version: HTTP/1.1
