+++ test/CacheCow.Client.Tests/packages.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.AspNet.WebApi.Client" version="4.0.20710.0" targetFramework="net40" />
+  <package id="Microsoft.Net.Http" version="2.0.20710.0" targetFramework="net40" />
+  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net40" />
+  <package id="NUnit" version="2.6.2" targetFramework="net40" />
+  <package id="RhinoMocks" version="3.6.1" targetFramework="net40" />
+</packages>
